refactor: move schedule test to schedule test file
