+++ tests/aws/services/events/test_events_schedule.py
@@ -5,6 +5,7 @@ from datetime import timedelta, timezone
 import pytest
 from botocore.exceptions import ClientError
 
+from localstack.aws.api.lambda_ import Runtime
 from localstack.testing.aws.eventbus_utils import trigger_scheduled_rule
 from localstack.testing.pytest import markers
 from localstack.testing.snapshots.transformer_utility import TransformerUtility
@@ -13,8 +14,10 @@ from localstack.utils.sync import retry
 from tests.aws.services.events.helper_functions import (
     events_time_string_to_timestamp,
     get_cron_expression,
+    is_v2_provider,
     sqs_collect_messages,
 )
+from tests.aws.services.lambda_.test_lambda import TEST_LAMBDA_PYTHON_ECHO
 
 
 class TestScheduleRate:
@@ -266,6 +269,61 @@ class TestScheduleRate:
         events.sort(key=lambda event: event["timestamp"])
 
         snapshot.match("log-events", events)
+
+    @markers.aws.validated
+    @pytest.mark.skipif(is_v2_provider(), reason="V2 provider does not support this feature yet")
+    def test_should_ignore_schedules_for_put_event(
+        self, create_lambda_function, cleanups, aws_client
+    ):
+        """Regression test for https://github.com/localstack/localstack/issues/7847"""
+        fn_name = f"test-event-fn-{short_uid()}"
+        create_lambda_function(
+            func_name=fn_name,
+            handler_file=TEST_LAMBDA_PYTHON_ECHO,
+            runtime=Runtime.python3_9,
+            client=aws_client.lambda_,
+        )
+
+        aws_client.lambda_.add_permission(
+            FunctionName=fn_name,
+            StatementId="AllowFnInvokeStatement",
+            Action="lambda:InvokeFunction",
+            Principal="events.amazonaws.com",
+        )
+
+        fn_arn = aws_client.lambda_.get_function(FunctionName=fn_name)["Configuration"][
+            "FunctionArn"
+        ]
+        aws_client.events.put_rule(
+            Name="ScheduledLambda", ScheduleExpression="rate(1 minute)"
+        )  # every minute, can't go lower than that
+        cleanups.append(lambda: aws_client.events.delete_rule(Name="ScheduledLambda"))
+        aws_client.events.put_targets(
+            Rule="ScheduledLambda", Targets=[{"Id": "calllambda1", "Arn": fn_arn}]
+        )
+        cleanups.append(
+            lambda: aws_client.events.remove_targets(Rule="ScheduledLambda", Ids=["calllambda1"])
+        )
+
+        aws_client.events.put_events(
+            Entries=[
+                {
+                    "Source": "MySource",
+                    "DetailType": "CustomType",
+                    "Detail": json.dumps({"message": "manually invoked"}),
+                }
+            ]
+        )
+
+        def check_invocation():
+            events_after = aws_client.logs.filter_log_events(logGroupName=f"/aws/lambda/{fn_name}")
+            # the custom sent event should NOT trigger the lambda (!)
+            assert len([e for e in events_after["events"] if "START" in e["message"]]) >= 1
+            assert (
+                len([e for e in events_after["events"] if "manually invoked" in e["message"]]) == 0
+            )
+
+        retry(check_invocation, sleep=5, retries=15)
 
 
 class TestScheduleCron:
