+++ tests/aws/services/events/test_events_targets.validation.json
@@ -10,8 +10,5 @@
   },
   "tests/aws/services/events/test_events_targets.py::TestEventTargetLambda::test_put_events_with_target_lambda_list_entry": {
     "last_validated_date": "2024-04-08T17:33:44+00:00"
-  },
-  "tests/aws/services/events/test_events_targets.py::test_should_ignore_schedules_for_put_event": {
-    "last_validated_date": "2024-03-26T15:51:47+00:00"
   }
 }
